Rename 'notification' module to 'notifications'
